# dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ApiFailResponse": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"FAIL"
						],
						"nullable": false
					},
					"error": {
						"properties": {
							"data": {},
							"reason": {
								"type": "string"
							},
							"errorCode": {
								"type": "string"
							}
						},
						"required": [
							"reason",
							"errorCode"
						],
						"type": "object"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ReviewData": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "* 리뷰 고유 식별자 ID",
						"example": 1
					},
					"content": {
						"type": "string",
						"description": "* 리뷰 본문 내용",
						"example": "음식이 정말 맛있고 배달도 빨라요! 추천합니다."
					},
					"score": {
						"type": "number",
						"format": "double",
						"description": "* 평점 점수 (예: 1점 ~ 5점)",
						"example": 5
					},
					"createdAt": {
						"type": "string",
						"description": "* 리뷰 작성 일시 (ISO 8601 string 또는 YYYY-MM-DD 형식)",
						"example": "2026-05-16T12:34:56.789Z"
					},
					"nickname": {
						"type": "string",
						"description": "* 리뷰를 작성한 유저의 닉네임",
						"example": "맛있는거좋아"
					},
					"reviewImages": {
						"items": {},
						"type": "array",
						"description": "* 리뷰에 첨부된 이미지 목록",
						"example": [
							"https://example.com/images/review1.jpg"
						]
					},
					"ownerReply": {
						"type": "string",
						"nullable": true,
						"description": "* 사장님이 작성한 답글 내용 (답글이 없으면 null)",
						"example": "소중한 리뷰 감사합니다! 다음에 또 주문해 주세요."
					}
				},
				"required": [
					"id",
					"content",
					"score",
					"createdAt",
					"nickname",
					"reviewImages",
					"ownerReply"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewListResponse": {
				"properties": {
					"reviewData": {
						"items": {
							"$ref": "#/components/schemas/ReviewData"
						},
						"type": "array",
						"description": "* 페이징 처리되어 반환된 리뷰 목록 데이터 배열"
					},
					"cursor": {
						"type": "number",
						"format": "double",
						"nullable": true,
						"description": "* 다음 페이지 조회를 위한 커서 ID (마지막 페이지라면 null)",
						"example": 15
					},
					"isLastPage": {
						"type": "boolean",
						"description": "* 다음 페이지가 없는 마지막 페이지인지 여부",
						"example": false
					}
				},
				"required": [
					"reviewData",
					"cursor",
					"isLastPage"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_ReviewListResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/ReviewListResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "playground-umc-10th-nodejs",
		"version": "1.0.0",
		"description": "UMC 테스트 프로젝트",
		"contact": {}
	},
	"paths": {
...
		"/reviews/stores/{storeId}": {
			"get": {
				"operationId": "HandleListStoreReviews",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ReviewListResponse_"
								}
							}
						}
					},
					"400": {
						"description": "유효하지 않은 사용자 정보 에러",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "400",
												"reason": "사용자 정보가 없습니다."
											},
											"data": null
										}
									}
								}
							}
						}
					}
				},
				"description": "특정 가게의 리뷰 목록 조회 API",
				"summary": "가게 ID를 기반으로 작성된 리뷰들을 페이징 처리하여 조회합니다.",
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": [
					{
						"description": "리뷰를 조회할 가게의 고유 식별자 ID",
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "한 번에 가져올 리뷰의 개수 (기본값: 10개)",
						"in": "query",
						"name": "take",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "페이징 처리를 위한 마지막 리뷰의 ID (커서 방식)",
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/reviews/me": {
			"get": {
				"operationId": "HandleListUserReviews",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ReviewListResponse_"
								}
							}
						}
					},
					"400": {
						"description": "유효하지 않은 사용자 정보 에러",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"error": {
												"errorCode": "400",
												"reason": "사용자 정보가 없습니다."
											},
											"data": null
										}
									}
								}
							}
						}
					}
				},
				"description": "내가 작성한 리뷰 목록 조회 API",
				"summary": "현재 로그인한 유저 본인이 작성한 리뷰들을 페이징 처리하여 조회합니다.",
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": [
					{
						"description": "한 번에 가져올 리뷰의 개수 (기본값: 10개)",
						"in": "query",
						"name": "take",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "페이징 처리를 위한 마지막 리뷰의 ID (커서 방식)",
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}